Sync bitbucket and GitHub

+++ plugins/modules/na_ontap_user_role.py
@@ -193,6 +193,7 @@ class NetAppOntapUserRole:
                                 ('query', 'privileges')]
         )
         self.owner_uuid = None
+        self.role_modified = None
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
         if self.parameters.get('privileges') is not None:
@@ -391,6 +392,7 @@ class NetAppOntapUserRole:
 
     def modify_role_rest(self, modify):
         # there is no direct modify for role.
+        self.role_modified = False
         privileges = self.get_role_privileges_rest()
         modify_privilege = []
         for privilege in modify['privileges']:
@@ -399,16 +401,23 @@ class NetAppOntapUserRole:
             # if the path is not in privilege then it need to be added
             if path not in privileges:
                 self.create_role_privilege(privilege)
+                self.role_modified = True
             elif privilege.get('query'):
                 if not privileges[path].get('query'):
                     self.modify_role_privilege(privilege, path)
+                    self.role_modified = True
                 elif privilege['query'] != privileges[path]['query']:
                     self.modify_role_privilege(privilege, path)
+                    self.role_modified = True
             elif privilege.get('access') and privilege['access'] != privileges[path]['access']:
                 self.modify_role_privilege(privilege, path)
+                self.role_modified = True
         for privilege_path in privileges:
+            if privilege_path == 'DEFAULT':
+                continue
             if privilege_path not in modify_privilege:
                 self.delete_role_privilege(privilege_path)
+                self.role_modified = True
 
     def get_role_privileges_rest(self):
         api = 'security/roles/%s/%s/privileges' % (self.owner_uuid, self.parameters['name'])
@@ -490,7 +499,11 @@ class NetAppOntapUserRole:
             elif modify:
                 self.modify_role(modify)
             if self.use_rest:
-                self.validate_create_modify_required(current, modify)
+                if self.role_modified is False:
+                    self.na_helper.changed = False
+                    modify = {}
+                else:
+                    self.validate_create_modify_required(current, modify)
         result = netapp_utils.generate_result(self.na_helper.changed, cd_action, modify)
         self.module.exit_json(**result)
